Extract inventory item rendering from the interaction
